Fixed get_options and get_option_name/value
