+++ admin/settings/page-connections.php
@@ -6,6 +6,7 @@ defined('WPINC') || die;
 class ConnectionPage extends SettingsPage
 {
     protected $settings_path = 'lnp_settings_connections';
+    protected $template_html = 'settings/page-connections.php';
     protected $option_name = 'lnp_connection';
 
     protected $page_title = 'Wallet Settings';
@@ -241,12 +242,12 @@ class ConnectionPage extends SettingsPage
     public function field_connection_types()
     {
         echo '<div class="wp-lnp-card__container">';
-        $this->connection_card('lnd', $this->plugin->get_file_url('img/lnd.png'), 'Lnd', 'Connect using lndhub');
-        $this->connection_card('lndhub', $this->plugin->get_file_url('img/lndhub.png'), 'LndHub (BlueWallet)', 'Connect using lndhub');
-        $this->connection_card('lnbits', $this->plugin->get_file_url('img/lnbits.png'), 'LNbits', 'Connect to your LNbits account');
-        $this->connection_card('lnaddress', $this->plugin->get_file_url('img/satsymbol-black.png'), 'Lightning Address Config', 'Connect using Lightning Address Config');
-        $this->connection_card('btcpay', $this->plugin->get_file_url('img/BTCPay_Icon_with_background.png'), 'BTC Pay', 'Connect using BTCPay Server');
-        $this->connection_card('lndhub_create_account', $this->plugin->get_file_url('img/alby.png'), 'Create a new wallet', 'We create and manage a lightning wallet for you');
+        $this->connection_card('lnd', $this->plugin->get_file_url('assets/img/lnd.png'), 'Lnd', 'Connect using lndhub');
+        $this->connection_card('lndhub', $this->plugin->get_file_url('assets/img/lndhub.png'), 'LndHub (BlueWallet)', 'Connect using lndhub');
+        $this->connection_card('lnbits', $this->plugin->get_file_url('assets/img/lnbits.png'), 'LNbits', 'Connect to your LNbits account');
+        $this->connection_card('lnaddress', $this->plugin->get_file_url('assets/img/satsymbol-black.png'), 'Lightning Address Config', 'Connect using Lightning Address Config');
+        $this->connection_card('btcpay', $this->plugin->get_file_url('assets/img/BTCPay_Icon_with_background.png'), 'BTC Pay', 'Connect using BTCPay Server');
+        $this->connection_card('lndhub_create_account', $this->plugin->get_file_url('assets/img/alby.png'), 'Create a new wallet', 'We create and manage a lightning wallet for you');
         echo '</div>';
     }
 
@@ -347,115 +348,5 @@ class ConnectionPage extends SettingsPage
         foreach ((array) $wp_settings_fields[$page][$section] as $field) {
             call_user_func($field['callback'], $field['args']);
         }
-    }
-
-    public function renderer()
-    {
-
-?>
-        <div class="wrap">
-            <h1>Lightning Wallet Settings</h1>
-            <div class="node-info">
-                <?php
-                try {
-
-                    if ($this->plugin->getLightningClient() && $this->plugin->getLightningClient()->isConnectionValid()) {
-                        $node_info = $this->plugin->getLightningClient()->getInfo();
-                        echo "Connected to: " . $node_info['alias'] . ' - ' . $node_info['identity_pubkey'];
-                    } else {
-                        echo 'Not connected';
-                    }
-                } catch (Exception $e) {
-                    echo "Failed to connect: " . $e;
-                }
-                ?>
-            </div>
-            <form id="wallet_settings_form" method="post" action="options.php">
-                <script>
-                    let generating_lndhub_account = false;
-
-                    function lndhubCreateAccount(e) {
-                        const button = e.target;
-
-                        if (generating_lndhub_account) return;
-
-                        generating_lndhub_account = true;
-
-                        button.disabled = true;
-                        button.innerHTML = "Generating.....";
-                        const data = new FormData();
-
-                        data.append('action', 'create_lnp_hub_account');
-
-                        fetch("<?php echo get_site_url(); ?>/wp-admin/admin-ajax.php", {
-                                method: "POST",
-                                credentials: 'same-origin',
-                                body: data
-                            })
-                            .then((response) => response.json())
-                            .then((response) => {
-                                document.getElementById('lndhub_url').value = response.url;
-                                document.getElementById('lndhub_login').value = response.login;
-                                document.getElementById('lndhub_password').value = response.password;
-                                button.innerHTML = "Generated";
-                                document.getElementById("submit").click();
-                            }).catch((e) => console.error(e))
-                    }
-
-                    window.addEventListener("DOMContentLoaded", function() {
-                        document.getElementById('load_from_lndconnect').addEventListener('click', function(e) {
-                            e.preventDefault();
-                            var lndConnectUrl = prompt('Please enter your lndconnect string (e.g. run: lndconnect --url --port=8080)');
-                            if (!lndConnectUrl) {
-                                return;
-                            }
-                            var url = new URL(lndConnectUrl);
-                            document.getElementById('lnp_lnd_address').value = 'https:' + url.pathname;
-                            document.getElementById('lnp_lnd_macaroon').value = url.searchParams.get('macaroon');
-                            document.getElementById('lnp_lnd_cert').value = url.searchParams.get('cert');
-                        });
-
-                        const list = document.querySelectorAll('.wp-lnp-card');
-                        const sections = document.querySelectorAll('.wp-lnp-section__hidden');
-                        const fieldsContainer = document.getElementById('wp-lnp-fields');
-                        const sectionsContent = {};
-                        sections.forEach(section => {
-                            sectionsContent[section.id] = section.innerHTML;
-                            section.remove();
-                        });
-                        const submitButton = document.getElementById('submit');
-                        list.forEach(el => el.addEventListener('click', e => {
-                            e.preventDefault();
-                            const current = el.attributes['data-section'].value;
-                            fieldsContainer.innerHTML = sectionsContent[current];
-                            if (current === 'lndhub_create_account') {
-                                submitButton.style.display = 'none';
-                                document.getElementById("create_lndhub_account").addEventListener("click", lndhubCreateAccount);
-                                return;
-                            }
-                            document.getElementById("create_lndhub_account").removeEventListener("click", lndhubCreateAccount);
-                            submitButton.style.display = 'block';
-                        }));
-
-
-                    });
-                </script>
-                <?php
-                settings_fields($this->settings_path);
-
-                echo '<div id="wp-lnp-fields">';
-
-                $this->render_current_section($this->settings_path);
-
-                echo '</div>';
-                submit_button();
-                ?>
-            </form>
-            <?php
-            $this->do_settings_sections($this->settings_path);
-            $this->do_settings_fields($this->settings_path, 'connection-types');
-            ?>
-        </div>
-<?php
     }
 }
